Run Jitter Analysis

Run GMACS jitter analysis, then save and plot results.

gmacs_do_jitter(
  gmacs.dat,
  sd,
  iter,
  wait = T,
  save_csv = T,
  csv_dir = NULL,
  save_plot = T,
  plot_dir = NULL,
  plot_only = F,
  model_name = NULL,
  version = NULL
)

Arguments

gmacs.dat

File path to gmacs.dat file.

sd

Jitter standard deviation.

iter

Number of jittering runs.

wait

Passed to shell(): a logical (not NA) indicating whether the R interpreter should wait for the command to finish, or run it asynchronously. Default = T.

save_csv

T/F save csv file output. Default = T.

csv_dir

Null. Directory in which to save output. If NULL, a directory called 'output' will be created in the same directory as gmacs.dat.

save_plot

T/F save plot. Default = T.

plot_dir

Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.

plot_only

T/F only plot results (analysis already run). Default = F.

model_name

NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".

version

NULL. Character string denoting GMACS version. Default: "2.20.16".

Value

A subdirectory called 'jitter' will be created in the same directory as the gmacs.dat file. Within './jitter' a subdirectory corresponding to each run will be created. The output and plot will be saved in the provided directories or output as a list if save_csv and save_plot = F.

Examples

gmacs_do_jitter(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat", sd = 0.3, iter = 100)
#> Error in gmacs_do_jitter(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat",     sd = 0.3, iter = 100): Cannot find gmacs.exe!!